+++ packages/svelte-query-devtools/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/svelte-query-devtools
 
+## 6.1.28
+
+### Patch Changes
+
+- Updated dependencies [[`3d21cac`](https://github.com/TanStack/query/commit/3d21cacdec3028b700c4c2e3e0ff8dbe7a235e8c)]:
+  - @tanstack/query-devtools@5.100.9
+  - @tanstack/svelte-query@6.1.28
+
 ## 6.1.27
 
 ### Patch Changes